pr36822.c: Avoid cast to pointer from integer of different size warning with -m31.
	* gcc.target/s390/pr36822.c: Avoid cast to pointer from integer
	of different size warning with -m31.

From-SVN: r141462

+++ gcc/testsuite/gcc.target/s390/pr36822.c
@@ -9,7 +9,7 @@ int boo()
 {
   struct {
     unsigned char pad[4096];
-    unsigned long long bar;
+    unsigned long bar;
   } *foo;
   asm volatile( "" : "=m" (*(unsigned long long*)(foo->bar))
 		: "a" (&foo->bar));
